fix migraciones

+++ database/migrations/2020_02_27_191019_prestacion_presupuesto_table.php
@@ -16,7 +16,7 @@ class PrestacionPresupuestoTable extends Migration
         Schema::create('prestacion_presupuesto', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->unsignedBigInteger('prestacion_id')->nullable();
-            $table->unsignedInteger('presupuesto_id')->nullable();
+            $table->unsignedBigInteger('presupuesto_id')->nullable();
             $table->timestamps();
         });
